requirements.txt
argcomplete==3.5.0
PyYAML==6.0.2
setuptools==59.6.0
boto3==1.34.161
